Handle ip_output failures better
If ip_output fails for a tcp segment, the segment is still treated as
sent and moved to the unacknowledged list, where it won't get sent again
until we get to a retransmit path. Since that takes some time and has
other effects we'd rather treat it as unsent and try sending it again
later.